add WebSocket real-time market streaming (prices, trades, order book)

+++ src/main/java/com/trading/tradingplatform/service/UserService.java
@@ -11,6 +11,8 @@ import lombok.AllArgsConstructor;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
+
 @Service
 @AllArgsConstructor
 public class UserService {
@@ -48,7 +50,7 @@ public class UserService {
         user.setUsername(request.getUsername());
         user.setEmail(request.getEmail());
         user.setPassword(passwordEncoder.encode(request.getPassword()));
-        user.setBalance(1000.0);
+        user.setBalance(BigDecimal.valueOf(1000.0));
 
         return userRepository.save(user);
     }
